# docs/04_openapi.yaml
openapi: 3.0.3
info:
  title: my_race API
  version: 1.0.0
  description: 한국마사회 공공데이터 기반 경주 정보·추천 순위 API. 예측은 참고 정보입니다.
servers:
  - url: http://localhost:8080/v1
components:
  parameters:
    DeviceId:
      name: X-Device-Id
      in: header
      required: true
      schema: { type: string, format: uuid }
    Meet:
      name: meet
      in: query
      schema: { type: integer, enum: [1, 2, 3] }
      description: 1 서울, 2 제주, 3 부경
    Date:
      name: date
      in: query
      schema: { type: string, format: date }
  schemas:
    Error:
      type: object
      properties:
        code: { type: string }
        message: { type: string }
        traceId: { type: string }
    TopPick:
      type: object
      properties:
        gateNo: { type: integer }
        horseName: { type: string }
        hrNo: { type: string }
        score: { type: number }
    RaceSummary:
      type: object
      properties:
        id: { type: integer, format: int64 }
        raceNo: { type: integer }
        distance: { type: integer }
        track: { type: string }
        grade: { type: string }
        startTime: { type: string }
        status: { type: string, enum: [PLANNED, ENTRIES_FIXED, FINISHED, CANCELED] }
        entryCount: { type: integer }
        topPick: { $ref: '#/components/schemas/TopPick', nullable: true }
    EntryResult:
      type: object
      properties:
        finishPos: { type: integer }
        finishTime: { type: string }
        winOdds: { type: number }
        placeOdds: { type: number }
    EntryView:
      type: object
      properties:
        rank: { type: integer }
        gateNo: { type: integer }
        hrNo: { type: string }
        horseName: { type: string }
        jkNo: { type: string }
        jockeyName: { type: string }
        trNo: { type: string }
        trainerName: { type: string }
        weightCarried: { type: number }
        horseWeight: { type: number, nullable: true }
        score: { type: number, nullable: true }
        factors:
          type: object
          nullable: true
          additionalProperties: { type: string }
        locked: { type: boolean }
        result: { $ref: '#/components/schemas/EntryResult', nullable: true }
    RaceDetail:
      type: object
      properties:
        id: { type: integer, format: int64 }
        meet: { type: integer }
        meetName: { type: string }
        date: { type: string, format: date }
        raceNo: { type: integer }
        distance: { type: integer }
        track: { type: string }
        grade: { type: string }
        startTime: { type: string }
        status: { type: string }
        engineVersion: { type: string, nullable: true }
        computedAt: { type: string, format: date-time, nullable: true }
        premium: { type: boolean }
        freeRankLimit: { type: integer }
        entries:
          type: array
          items: { $ref: '#/components/schemas/EntryView' }
        disclaimer: { type: string }
    HitFlags:
      type: object
      properties:
        winHit: { type: boolean }
        placeHit: { type: boolean }
        showHit: { type: boolean }
    HitSummary:
      type: object
      properties:
        from: { type: string, format: date }
        to: { type: string, format: date }
        engineVersion: { type: string }
        races: { type: integer }
        winRate: { type: number }
        placeRate: { type: number }
        showRate: { type: number }
        byMeet:
          type: array
          items:
            type: object
            properties:
              meet: { type: integer }
              meetName: { type: string }
              races: { type: integer }
              winRate: { type: number }
        weekly:
          type: array
          items:
            type: object
            properties:
              weekStart: { type: string, format: date }
              races: { type: integer }
              winRate: { type: number }
        detailLocked: { type: boolean }
paths:
  /devices:
    post:
      summary: 기기 등록/갱신
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [deviceId, platform, adultConfirmed]
              properties:
                deviceId: { type: string }
                platform: { type: string, enum: [ANDROID, IOS] }
                adultConfirmed: { type: boolean }
                preferredMeet: { type: integer }
                pushToken: { type: string }
                prefs:
                  type: object
                  properties:
                    notifyEntries: { type: boolean }
                    notifyRankUpdate: { type: boolean }
                    notifyPreRace: { type: boolean }
      responses:
        '200': { description: OK }
  /devices/me:
    get:
      parameters: [ { $ref: '#/components/parameters/DeviceId' } ]
      responses:
        '200': { description: OK }
  /races:
    get:
      parameters:
        - { $ref: '#/components/parameters/DeviceId' }
        - { $ref: '#/components/parameters/Meet' }
        - { $ref: '#/components/parameters/Date' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  meet: { type: integer, nullable: true }
                  date: { type: string, format: date }
                  updatedAt: { type: string, format: date-time }
                  races:
                    type: array
                    items: { $ref: '#/components/schemas/RaceSummary' }
  /races/{id}:
    get:
      parameters:
        - { $ref: '#/components/parameters/DeviceId' }
        - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RaceDetail' }
        '404':
          description: not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /results:
    get:
      parameters:
        - { $ref: '#/components/parameters/DeviceId' }
        - { $ref: '#/components/parameters/Meet' }
        - { $ref: '#/components/parameters/Date' }
      responses:
        '200': { description: OK }
  /hits/summary:
    get:
      parameters:
        - { $ref: '#/components/parameters/DeviceId' }
        - { $ref: '#/components/parameters/Meet' }
        - { name: weeks, in: query, schema: { type: integer, default: 4 } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/HitSummary' }
  /horses/{hrNo}:
    get:
      parameters:
        - { $ref: '#/components/parameters/DeviceId' }
        - { name: hrNo, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: OK }
  /jockeys/{jkNo}:
    get:
      parameters:
        - { $ref: '#/components/parameters/DeviceId' }
        - { name: jkNo, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: OK }
  /trainers/{trNo}:
    get:
      parameters:
        - { $ref: '#/components/parameters/DeviceId' }
        - { name: trNo, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: OK }
  /subscriptions/verify:
    post:
      parameters: [ { $ref: '#/components/parameters/DeviceId' } ]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                platform: { type: string }
                productId: { type: string }
                purchaseToken: { type: string }
      responses:
        '200': { description: OK }